move GetDataTypeName to problem_descrption_base.hpp, organize includes
