+++ src/include/miopen/activ/problem_description.hpp
@@ -27,8 +27,8 @@
 #pragma once
 
 #include <miopen/problem_description_base.hpp>
-#include <miopen/activ.hpp>
 #include <miopen/tensor.hpp>
+#include <miopen/activ.hpp>
 
 #include <string>
